+++ packages/firebase_database/ios/Classes/FirebaseDatabasePlugin.m
@@ -30,15 +30,15 @@
 }
 @end
 
-FIRDatabaseReference *getReference(NSDictionary *arguments) {
+FIRDatabaseReference *getReference(FIRDatabase *database, NSDictionary *arguments) {
   NSString *path = arguments[@"path"];
-  FIRDatabaseReference *ref = [FIRDatabase database].reference;
+  FIRDatabaseReference *ref = database.reference;
   if ([path length] > 0) ref = [ref child:path];
   return ref;
 }
 
-FIRDatabaseQuery *getQuery(NSDictionary *arguments) {
-  FIRDatabaseQuery *query = getReference(arguments);
+FIRDatabaseQuery *getQuery(FIRDatabase *database, NSDictionary *arguments) {
+  FIRDatabaseQuery *query = getReference(database, arguments);
   NSDictionary *parameters = arguments[@"parameters"];
   NSString *orderBy = parameters[@"orderBy"];
   if ([orderBy isEqualToString:@"child"]) {
@@ -154,23 +154,30 @@ id roundDoubles(id value) {
 }
 
 - (void)handleMethodCall:(FlutterMethodCall *)call result:(FlutterResult)result {
+  FIRDatabase *database;
+  NSString *appName = call.arguments[@"app"];
+  if (![appName isEqual:[NSNull null]]) {
+    database = [FIRDatabase databaseForApp:[FIRApp appNamed:appName]];
+  } else {
+    database = [FIRDatabase database];
+  }
   void (^defaultCompletionBlock)(NSError *, FIRDatabaseReference *) =
       ^(NSError *error, FIRDatabaseReference *ref) {
         result(error.flutterError);
       };
   if ([@"FirebaseDatabase#goOnline" isEqualToString:call.method]) {
-    [[FIRDatabase database] goOnline];
+    [database goOnline];
     result(nil);
   } else if ([@"FirebaseDatabase#goOffline" isEqualToString:call.method]) {
-    [[FIRDatabase database] goOffline];
+    [database goOffline];
     result(nil);
   } else if ([@"FirebaseDatabase#purgeOutstandingWrites" isEqualToString:call.method]) {
-    [[FIRDatabase database] purgeOutstandingWrites];
+    [database purgeOutstandingWrites];
     result(nil);
   } else if ([@"FirebaseDatabase#setPersistenceEnabled" isEqualToString:call.method]) {
-    NSNumber *value = call.arguments;
+    NSNumber *value = call.arguments[@"enabled"];
     @try {
-      [FIRDatabase database].persistenceEnabled = value.boolValue;
+      database.persistenceEnabled = value.boolValue;
       result([NSNumber numberWithBool:YES]);
     } @catch (NSException *exception) {
       if ([@"FIRDatabaseAlreadyInUse" isEqualToString:exception.name]) {
@@ -181,9 +188,9 @@ id roundDoubles(id value) {
       }
     }
   } else if ([@"FirebaseDatabase#setPersistenceCacheSizeBytes" isEqualToString:call.method]) {
-    NSNumber *value = call.arguments;
+    NSNumber *value = call.arguments[@"cacheSize"];
     @try {
-      [FIRDatabase database].persistenceCacheSizeBytes = value.unsignedIntegerValue;
+      database.persistenceCacheSizeBytes = value.unsignedIntegerValue;
       result([NSNumber numberWithBool:YES]);
     } @catch (NSException *exception) {
       if ([@"FIRDatabaseAlreadyInUse" isEqualToString:exception.name]) {
@@ -194,18 +201,18 @@ id roundDoubles(id value) {
       }
     }
   } else if ([@"DatabaseReference#set" isEqualToString:call.method]) {
-    [getReference(call.arguments) setValue:call.arguments[@"value"]
-                               andPriority:call.arguments[@"priority"]
-                       withCompletionBlock:defaultCompletionBlock];
+    [getReference(database, call.arguments) setValue:call.arguments[@"value"]
+                                         andPriority:call.arguments[@"priority"]
+                                 withCompletionBlock:defaultCompletionBlock];
   } else if ([@"DatabaseReference#update" isEqualToString:call.method]) {
-    [getReference(call.arguments) updateChildValues:call.arguments[@"value"]
-                                withCompletionBlock:defaultCompletionBlock];
+    [getReference(database, call.arguments) updateChildValues:call.arguments[@"value"]
+                                          withCompletionBlock:defaultCompletionBlock];
   } else if ([@"DatabaseReference#setPriority" isEqualToString:call.method]) {
-    [getReference(call.arguments) setPriority:call.arguments[@"priority"]
-                          withCompletionBlock:defaultCompletionBlock];
+    [getReference(database, call.arguments) setPriority:call.arguments[@"priority"]
+                                    withCompletionBlock:defaultCompletionBlock];
   } else if ([@"DatabaseReference#runTransaction" isEqualToString:call.method]) {
-    [getReference(call.arguments) runTransactionBlock:^FIRTransactionResult *_Nonnull(
-                                      FIRMutableData *_Nonnull currentData) {
+    [getReference(database, call.arguments) runTransactionBlock:^FIRTransactionResult *_Nonnull(
+                                                FIRMutableData *_Nonnull currentData) {
       // Create semaphore to allow native side to wait while snapshot
       // updates occur on the Dart side.
       dispatch_semaphore_t semaphore = dispatch_semaphore_create(0);
@@ -249,7 +256,7 @@ id roundDoubles(id value) {
             [self.updatedSnapshots objectForKey:call.arguments[@"transactionKey"]][@"value"];
       } else {
         if (result != 0) {
-          NSLog(@"Transaction at %@ timed out.", [getReference(call.arguments) URL]);
+          NSLog(@"Transaction at %@ timed out.", [getReference(database, call.arguments) URL]);
         }
         return [FIRTransactionResult abort];
       }
@@ -268,7 +275,8 @@ id roundDoubles(id value) {
         }];
   } else if ([@"Query#observe" isEqualToString:call.method]) {
     FIRDataEventType eventType = parseEventType(call.arguments[@"eventType"]);
-    __block FIRDatabaseHandle handle = [getQuery(call.arguments) observeEventType:eventType
+    __block FIRDatabaseHandle handle = [getQuery(database, call.arguments)
+        observeEventType:eventType
         andPreviousSiblingKeyWithBlock:^(FIRDataSnapshot *snapshot, NSString *previousSiblingKey) {
           [self.channel invokeMethod:@"Event"
                            arguments:@{
@@ -290,11 +298,11 @@ id roundDoubles(id value) {
     result([NSNumber numberWithUnsignedInteger:handle]);
   } else if ([@"Query#removeObserver" isEqualToString:call.method]) {
     FIRDatabaseHandle handle = [call.arguments[@"handle"] unsignedIntegerValue];
-    [getQuery(call.arguments) removeObserverWithHandle:handle];
+    [getQuery(database, call.arguments) removeObserverWithHandle:handle];
     result(nil);
   } else if ([@"Query#keepSynced" isEqualToString:call.method]) {
     NSNumber *value = call.arguments[@"value"];
-    [getQuery(call.arguments) keepSynced:value.boolValue];
+    [getQuery(database, call.arguments) keepSynced:value.boolValue];
     result(nil);
   } else {
     result(FlutterMethodNotImplemented);