+++ packages/firebase_database/test/firebase_database_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:async';
 
+import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -16,7 +17,15 @@ void main() {
 
     int mockHandleId = 0;
     final List<MethodCall> log = <MethodCall>[];
-    final FirebaseDatabase database = FirebaseDatabase.instance;
+    final FirebaseApp app = const FirebaseApp(
+      name: 'testApp',
+      options: const FirebaseOptions(
+        googleAppID: '1:1234567890:ios:42424242424242',
+        gcmSenderID: '1234567890',
+        databaseURL: 'https://fake-database-url.firebaseio.com',
+      ),
+    );
+    final FirebaseDatabase database = new FirebaseDatabase(app: app);
 
     setUp(() async {
       channel.setMockMethodCallHandler((MethodCall methodCall) async {
@@ -82,11 +91,17 @@ void main() {
         <Matcher>[
           isMethodCall(
             'FirebaseDatabase#setPersistenceEnabled',
-            arguments: false,
+            arguments: <String, dynamic>{
+              'app': app.name,
+              'enabled': false,
+            },
           ),
           isMethodCall(
             'FirebaseDatabase#setPersistenceEnabled',
-            arguments: true,
+            arguments: <String, dynamic>{
+              'app': app.name,
+              'enabled': true,
+            },
           ),
         ],
       );
@@ -99,7 +114,10 @@ void main() {
         <Matcher>[
           isMethodCall(
             'FirebaseDatabase#setPersistenceCacheSizeBytes',
-            arguments: 42,
+            arguments: <String, dynamic>{
+              'app': app.name,
+              'cacheSize': 42,
+            },
           ),
         ],
       );
@@ -110,7 +128,12 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('FirebaseDatabase#goOnline', arguments: null),
+          isMethodCall(
+            'FirebaseDatabase#goOnline',
+            arguments: <String, dynamic>{
+              'app': app.name,
+            },
+          ),
         ],
       );
     });
@@ -120,7 +143,12 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('FirebaseDatabase#goOffline', arguments: null),
+          isMethodCall(
+            'FirebaseDatabase#goOffline',
+            arguments: <String, dynamic>{
+              'app': app.name,
+            },
+          ),
         ],
       );
     });
@@ -132,7 +160,9 @@ void main() {
         <Matcher>[
           isMethodCall(
             'FirebaseDatabase#purgeOutstandingWrites',
-            arguments: null,
+            arguments: <String, dynamic>{
+              'app': app.name,
+            },
           ),
         ],
       );
@@ -150,6 +180,7 @@ void main() {
             isMethodCall(
               'DatabaseReference#set',
               arguments: <String, dynamic>{
+                'app': app.name,
                 'path': 'foo',
                 'value': value,
                 'priority': null,
@@ -158,6 +189,7 @@ void main() {
             isMethodCall(
               'DatabaseReference#set',
               arguments: <String, dynamic>{
+                'app': app.name,
                 'path': 'bar',
                 'value': value,
                 'priority': priority,
@@ -174,7 +206,11 @@ void main() {
           <Matcher>[
             isMethodCall(
               'DatabaseReference#update',
-              arguments: <String, dynamic>{'path': 'foo', 'value': value},
+              arguments: <String, dynamic>{
+                'app': app.name,
+                'path': 'foo',
+                'value': value,
+              },
             ),
           ],
         );
@@ -188,7 +224,11 @@ void main() {
           <Matcher>[
             isMethodCall(
               'DatabaseReference#setPriority',
-              arguments: <String, dynamic>{'path': 'foo', 'priority': priority},
+              arguments: <String, dynamic>{
+                'app': app.name,
+                'path': 'foo',
+                'priority': priority,
+              },
             ),
           ],
         );
@@ -211,6 +251,7 @@ void main() {
             isMethodCall(
               'DatabaseReference#runTransaction',
               arguments: <String, dynamic>{
+                'app': app.name,
                 'path': 'foo',
                 'transactionKey': 0,
                 'transactionTimeout': 5000,
@@ -243,6 +284,7 @@ void main() {
             isMethodCall(
               'Query#keepSynced',
               arguments: <String, dynamic>{
+                'app': app.name,
                 'path': path,
                 'parameters': <String, dynamic>{},
                 'value': true,
@@ -277,6 +319,7 @@ void main() {
             isMethodCall(
               'Query#keepSynced',
               arguments: <String, dynamic>{
+                'app': app.name,
                 'path': path,
                 'parameters': expectedParameters,
                 'value': false
@@ -373,6 +416,7 @@ void main() {
             isMethodCall(
               'Query#observe',
               arguments: <String, dynamic>{
+                'app': app.name,
                 'path': path,
                 'parameters': <String, dynamic>{},
                 'eventType': '_EventType.value',
@@ -381,6 +425,7 @@ void main() {
             isMethodCall(
               'Query#removeObserver',
               arguments: <String, dynamic>{
+                'app': app.name,
                 'path': path,
                 'parameters': <String, dynamic>{},
                 'handle': 87,