+++ packages/firebase_core/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Fixes for database URL on Android
+* Make GCM sender id optional on Android
+* Relax GMS dependency to 11.+
+
 ## 0.0.1
 
 * Initial Release