+++ packages/firebase_database/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* Support for multiple databases, new dependency on firebase_core
+* Relax GMS dependency to 11.+
+
 ## 0.1.4
 
 * Add FLT prefix to iOS types